Renamed controller and model. Added Stock method to V1Controller. Gave up on sql.
Renamed controller and model to be appropriate for the expectations. Created Stock method which is a takes a Post request and search for each value in the saved stock database. If the value is found then it updates the volume with the given one. If not then creates the value with the given volume. Deleted sql type project, because of too many obstacles.